Episode: Fixed an Issue with Downloading Non-CDN Videos
The issue basically was that non-cdn downloads would be prioritized even
though I seem to have deleted the code to download non-cdn videos.
Reasonably, I did not re-add the code to download non-cdn videos I just
changed it to prioritize cdn first. Surely this will not come back to
bite me in the ass.
Fixes issue #36 (this does actually fix it btw because it is now
possible to download the episode in the issue)